refactor: migrate experimental Megatron components to main API; fix: fix several bugs in FSDP engine
