+++ areal/engine/ppo/actor.py
@@ -6,6 +6,7 @@ import torch
 from areal.api.cli_args import MicroBatchSpec, PPOActorConfig
 from areal.api.engine_api import TrainEngine
 from areal.engine.fsdp_engine import FSDPEngine
+from areal.engine.megatron_engine import MegatronEngine
 from areal.utils import stats_tracker
 from areal.utils.data import (
     KLEstimator,
@@ -54,7 +55,10 @@ class PPOActor:
         temperature: Optional[float] = None,
     ) -> torch.Tensor | None:
         def calc_logprobs(logits, input_data):
-            labels = torch.roll(input_data["input_ids"], shifts=-1, dims=-1)
+            labels = input_data.get(
+                "rolled_input_ids",
+                torch.roll(input_data["input_ids"], shifts=-1, dims=-1),
+            )
             logprobs = gather_logprobs(logits, labels, temperature or 1.0)
             return logprobs
 
@@ -288,6 +292,24 @@ class FSDPPPOActor(FSDPEngine):
         return self.actor.ppo_update(*args, **kwargs)
 
 
+class MegatronPPOActor(MegatronEngine):
+
+    def __init__(self, config: PPOActorConfig):
+        super().__init__(config)
+        self.actor = PPOActor(config, self)
+
+    @torch.no_grad()
+    def compute_logp(self, *args, **kwargs) -> torch.Tensor | None:
+        return self.actor.compute_logp(*args, **kwargs)
+
+    @torch.no_grad()
+    def compute_advantages(self, *args, **kwargs) -> None:
+        self.actor.compute_advantages(*args, **kwargs)
+
+    def ppo_update(self, *args, **kwargs) -> List[Dict[str, float]]:
+        return self.actor.ppo_update(*args, **kwargs)
+
+
 def grpo_loss_fn(
     logits: torch.Tensor,
     input_data: Dict,
@@ -299,14 +321,14 @@ def grpo_loss_fn(
 ):
     """Loss function for actor step, all inputs should be splitted into
     pipeline micro batches, returns loss and logging stats."""
+    # Use rolled input_ids. Ulysses SP will roll input_ids in ulysses_prepare_inputs().
     labels = input_data.get(
         "rolled_input_ids",
         torch.roll(input_data["input_ids"], shifts=-1, dims=-1),
     )
     old_logp = input_data["logprobs"]
     advantages = input_data["advantages"]
-    # Use unsliced/full loss_mask.
-    # Ulysses SP will slice loss_mask in ulysses_prepare_inputs().
+    # Use full loss_mask. Ulysses SP will slice loss_mask in ulysses_prepare_inputs().
     loss_mask = input_data.get("full_loss_mask", input_data["loss_mask"]).bool()
     prox_logp = input_data["prox_logp"]
