+++ areal/engine/sft/lm_engine.py
@@ -5,6 +5,7 @@ import torch
 from areal.api.cli_args import TrainEngineConfig
 from areal.api.engine_api import TrainEngine
 from areal.engine.fsdp_engine import FSDPEngine
+from areal.engine.megatron_engine import MegatronEngine
 from areal.utils import stats_tracker
 from areal.utils.functional import gather_logprobs
 
@@ -42,14 +43,31 @@ class FSDPLMEngine(FSDPEngine):
         return self.lm_engine.evaluate_lm(data)
 
 
+class MegatronLMEngine(MegatronEngine):
+    def __init__(self, config: TrainEngineConfig):
+        super().__init__(config)
+        self.lm_engine = LMEngine(self)
+
+    def train_lm(self, data):
+        return self.lm_engine.train_lm(data)
+
+    def evaluate_lm(self, data):
+        return self.lm_engine.evaluate_lm(data)
+
+
 def compute_packed_sft_loss(
     logits: torch.Tensor, input_: Dict[str, Any]
 ) -> torch.Tensor:
-    packed_input_ids: torch.Tensor = input_["input_ids"]
+    # Use rolled input_ids. Ulysses SP will roll input_ids in ulysses_prepare_inputs().
+    labels: torch.Tensor = input_.get(
+        "rolled_input_ids",
+        torch.roll(input_["input_ids"], shifts=-1, dims=-1),
+    )
     cu_seqlens: torch.Tensor = input_["cu_seqlens"]
-    loss_mask = input_["loss_mask"].bool()
+    # Use full loss_mask. Ulysses SP will slice loss_mask in ulysses_prepare_inputs().
+    loss_mask = input_.get("full_loss_mask", input_["loss_mask"]).bool()
 
-    logprobs = gather_logprobs(logits, torch.roll(packed_input_ids, shifts=-1, dims=-1))
+    logprobs = gather_logprobs(logits, labels)
     loss_mask = torch.roll(loss_mask, shifts=-1, dims=-1)
     logprobs = torch.where(loss_mask, logprobs, 0)
