--- areal/experimental/megatron_lm_engine.py
@@ -1,102 +1,0 @@
-import torch
-from megatron.core import parallel_state as mpu
-from megatron.core import tensor_parallel
-
-from areal.api.engine_api import TrainEngine
-from areal.experimental.api.cli_args import (
-    ExperimentalTrainEngineConfig as TrainEngineConfig,
-)
-from areal.experimental.megatron_engine import MegatronEngine
-from areal.utils import stats_tracker
-from areal.utils.functional import gather_logprobs
-
-
-# TODO: merge this with areal/engine/sft/lm_engine.py when moving out of experimental.
-class LMEngine:
-    def __init__(self, engine: TrainEngine):
-        self.engine = engine
-
-    def train_lm(self, data):
-        self.engine.train()
-        return self.engine.train_batch(
-            input_=data,
-            loss_fn=compute_packed_sft_loss,
-            loss_weight_fn=lambda x: x["loss_mask"].count_nonzero(),
-        )
-
-    def evaluate_lm(self, data):
-        self.engine.eval()
-        return self.engine.eval_batch(
-            input_=data,
-            loss_fn=compute_packed_sft_loss,
-            loss_weight_fn=lambda x: x["loss_mask"].count_nonzero(),
-        )
-
-
-class MegatronLMEngine(MegatronEngine):
-    def __init__(self, config: TrainEngineConfig):
-        super().__init__(config)
-        self.lm_engine = LMEngine(self)
-
-    def train_lm(self, data):
-        return self.lm_engine.train_lm(data)
-
-    def evaluate_lm(self, data):
-        return self.lm_engine.evaluate_lm(data)
-
-
-def compute_packed_sft_loss(logits: torch.Tensor, input_) -> torch.Tensor:
-    packed_input_ids: torch.Tensor = input_["input_ids"]
-    cu_seqlens: torch.Tensor = input_["cu_seqlens"]
-    loss_mask = input_["loss_mask"].bool()
-
-    labels = torch.roll(packed_input_ids, shifts=-1, dims=-1)
-    if mpu.is_initialized() and mpu.get_tensor_model_parallel_world_size() > 1:
-        # NOTE: When tensor parallelism is enabled, logits are parallelized across TP ranks.
-        # If we explicitly gather logits, it will significantly increase GPU memory usage.
-        # Therefore here we use a utility function from megatron to avoid this.
-        logprobs = -tensor_parallel.vocab_parallel_cross_entropy(
-            vocab_parallel_logits=logits, target=labels
-        )
-    else:
-        logprobs = gather_logprobs(logits, labels)
-    loss_mask = torch.roll(loss_mask, shifts=-1, dims=-1)
-    logprobs = torch.where(loss_mask, logprobs, 0)
-
-    loss = -logprobs.sum() / loss_mask.count_nonzero()
-    with torch.no_grad():
-        seqlogp = torch.zeros(
-            cu_seqlens.shape[0] - 1, device=logits.device, dtype=torch.float64
-        )
-        for i in range(cu_seqlens.shape[0] - 1):
-            m = loss_mask[cu_seqlens[i] : cu_seqlens[i + 1]]
-            logp = logprobs[cu_seqlens[i] : cu_seqlens[i + 1]]
-            seqlogp[i] = torch.where(m, logp.detach(), 0.0).sum() / (m.count_nonzero())
-
-    ## Loggin stats
-    # NOTE: These stats are only available on pipeline last stage.
-    # Make sure to call stats_tracker.broadcast_stats() before exporting.
-    stats_tracker.denominator(
-        n_seqs=torch.ones(
-            cu_seqlens.shape[0] - 1, dtype=torch.bool, device=logprobs.device
-        ),
-        n_tokens=torch.ones(logits.shape[0], dtype=torch.bool, device=logits.device),
-        n_valid_tokens=loss_mask,
-        prompt_tokens=loss_mask.logical_not(),
-    )
-    stats_tracker.stat(
-        ppl=(-seqlogp).exp().float(),
-        denominator="n_seqs",
-    )
-    stats_tracker.stat(
-        loss=-logprobs.detach(),
-        denominator="n_valid_tokens",
-    )
-    vocab_min_logits = logits.detach().min(-1).values.float()
-    vocab_max_logits = logits.detach().max(-1).values.float()
-    stats_tracker.stat(
-        vocab_min_logits=vocab_min_logits,
-        vocab_max_logits=vocab_max_logits,
-        denominator="n_tokens",
-    )
-    return loss
