+++ areal/tests/test_megatron_engine.py
@@ -7,15 +7,13 @@ import torch
 from transformers import AutoTokenizer
 
 from areal.api.alloc_mode import AllocationMode
-from areal.api.io_struct import FinetuneSpec, SaveLoadMeta
-from areal.experimental.api.cli_args import (
-    ExperimentalTrainEngineConfig as TrainEngineConfig,
-)
-from areal.experimental.api.cli_args import (
+from areal.api.cli_args import (
     MegatronEngineConfig,
     OptimizerConfig,
+    TrainEngineConfig,
 )
-from areal.experimental.megatron_engine import MegatronEngine
+from areal.api.io_struct import FinetuneSpec, SaveLoadMeta
+from areal.engine.megatron_engine import MegatronEngine
 from areal.platforms import current_platform
 from areal.utils import logging
 from areal.utils.device import log_gpu_stats
