+++ areal/tests/torchrun/run_megatron_engine_distributed.py
@@ -10,17 +10,15 @@ from megatron.core import parallel_state as mpu
 from transformers import AutoTokenizer
 
 from areal.api.alloc_mode import AllocationMode
-from areal.api.io_struct import FinetuneSpec, SaveLoadMeta
-from areal.engine.fsdp_engine import FSDPEngine
-from areal.experimental.api.cli_args import (
-    ExperimentalTrainEngineConfig as TrainEngineConfig,
-)
-from areal.experimental.api.cli_args import (
+from areal.api.cli_args import (
     MegatronEngineConfig,
     MicroBatchSpec,
     OptimizerConfig,
+    TrainEngineConfig,
 )
-from areal.experimental.megatron_engine import MegatronEngine
+from areal.api.io_struct import FinetuneSpec, SaveLoadMeta
+from areal.engine.fsdp_engine import FSDPEngine
+from areal.engine.megatron_engine import MegatronEngine
 from areal.platforms import current_platform
 from areal.utils import seeding
 from areal.utils.data import broadcast_tensor_container
