+++ examples/math/gsm8k_sft_megatron.yaml
